% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flattenList.R
\name{flattenList}
\alias{flattenList}
\title{flattenList
flatten multi level list to simple list}
\usage{
flattenList(x)
}
\arguments{
\item{x}{a list with listed elements}
}
\value{
single level list
}
\description{
flattenList
flatten multi level list to simple list
}
