% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.caption.R
\name{get.caption}
\alias{get.caption}
\title{get.caption}
\usage{
get.caption(x, rm.html = TRUE, sentences = FALSE, letter.convert = TRUE)
}
\arguments{
\item{x}{A vector with HTML-coded tables.}

\item{rm.html}{logical. If TRUE, all HTML tags are removed, <sub> converts to '_', and <sup> to '^'.}

\item{sentences}{logical. If TRUE, a sentence vector is returned.}

\item{letter.convert}{logical. If TRUE, hexadecimal letters are converted to Unicode and unified with JATSdecoder::letter.convert.}
}
\value{
A character vector with the extracted caption text and NULL for no caption text
}
\description{
Extracts the content of HTML <caption>-tags.
}
