% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.footer.R
\name{get.footer}
\alias{get.footer}
\title{get.footer
Extracts the content of <table-wrap-footer>-tag.}
\usage{
get.footer(x, rm.html = TRUE, sentences = FALSE, letter.convert = TRUE)
}
\arguments{
\item{x}{A vector with HTML coded tables.}

\item{rm.html}{logical. If TRUE, all HTML tags are removed, <sub> converts to '_', <sup> to '^'.}

\item{sentences}{logical. If TRUE, a sentence vector is returned.}

\item{letter.convert}{logical. If TRUE, hexadecimal letters are converted to Unicode und unified with JATSdecoder::letter.convert.}
}
\value{
A character vector with the extracted footer text and NULL for no footer text
}
\description{
get.footer
Extracts the content of <table-wrap-footer>-tag.
}
