% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.tables.R
\name{get.tables}
\alias{get.tables}
\title{get.tables}
\usage{
get.tables(x)
}
\arguments{
\item{x}{HTML, HML, XML or CERMXML file or character object with HTML-encoded content.}
}
\value{
Character vector with one plain HTML-encoded table per cell.
}
\description{
Extracts HTML tables as vector of tables from plain HTML code, HTML, HML, XML or CERMXML files.
}
