% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessCaptionFooterDOCX.R
\name{guessCaptionFooterDOCX}
\alias{guessCaptionFooterDOCX}
\title{guessCaptionFooterDOCX
Extracts text blocks around tables within DOCX files in order to return the tables caption/footer.}
\usage{
guessCaptionFooterDOCX(x, MaxCaptionLength = 1, MaxFooterLength = 4)
}
\arguments{
\item{x}{character. A file path to a DOCX file.}

\item{MaxCaptionLength}{numeric. The maximum number of sentences within a text block that shall be treated as a caption. Text blocks that contain more sentences than this threshold are not extracted at all.}

\item{MaxFooterLength}{numeric. The maximum number of sentences within a text block that shall be treated as a footer. Text blocks that contain more sentences than this threshold are not extracted at all.}
}
\value{
A list with the extracted table captions and footers as vectors of length=number of tables.
}
\description{
guessCaptionFooterDOCX
Extracts text blocks around tables within DOCX files in order to return the tables caption/footer.
}
