% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2unicode.R
\name{html2unicode}
\alias{html2unicode}
\title{html2unicode}
\usage{
html2unicode(x)
}
\arguments{
\item{x}{A character vector or matrix.}
}
\value{
A character vector or matrix.
}
\description{
Converts HTML encoded special letters to unicode.
}
\examples{
html2unicode(x<-"&#34;, &#161;, &#162;.")
}
\references{
https://www.w3.org/TR/REC-html40/sgml/entities.html
}
