% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendCodings.R
\name{legendCodings}
\alias{legendCodings}
\title{legendCodings}
\usage{
legendCodings(x)
}
\arguments{
\item{x}{An HTML coded table or plain textual input.}
}
\value{
A list with detected p-value codings, abbreviations and sample size/s.
}
\description{
Extracts the coding of p-values, brackets, abbreviations, superscripts and reported sample size/s with N=number from tables caprion and footer notes/text.
}
