% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2text.R
\name{matrix2text}
\alias{matrix2text}
\title{matrix2text}
\usage{
matrix2text(
  x,
  legend = NULL,
  unifyMatrix = TRUE,
  expandAbbreviations = TRUE,
  standardPcoding = FALSE,
  rotate = FALSE,
  split = FALSE
)
}
\arguments{
\item{x}{A character matrix or list of character matrices.}

\item{legend}{A list with table legend codes extracted from table caption and/or footer with tableParser::legendCodings().}

\item{unifyMatrix}{Logical. If TRUE, matrix cells are unified for better post processing.}

\item{expandAbbreviations}{Logical. If TRUE, detected abbreviations are expanded to label detected in table caption/footer with tableParser::legendCodings().}

\item{standardPcoding}{Logical. If TRUE, and no other detection of p-value coding is detected, standard coding of p-values is assumed to be: * p<.05, ** p<.01 and *** p<.001.}

\item{rotate}{Logical. If TRUE, matrix/matrices is rotated before parsing.}

\item{split}{Logical. If TRUE, matrix/matrices are split for multi model tables.}
}
\value{
Character vector with a parsed and straight forward readable form of the input table. The result vector can be further processed with standardStats() to extract and structure the statistical standard test results only.
}
\description{
Convert character matrices to text.
}
\examples{
# some random data
x<-rnorm(100)
y<-x+rnorm(100)
# a model result table...
mod<-round(summary(lm(y~x))$coefficients,3)
rnames<-c("",rownames(mod))
cnames<-colnames(mod)
mod<-rbind(cnames,mod)
mod<-cbind(rnames,mod)
x<-unname(mod)
# ...as character result matrix
# parse matrix to text
matrix2text(x,unifyMatrix=FALSE)
}
