% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{prepareMatrix}
\usage{
prepareMatrix(x, split = FALSE)
}
\arguments{
\item{x}{character matrix}

\item{split}{logical. If TRUE multi model matrices are split to a list of single model matrices.}
}
\value{
character matrix
}
\description{
Prepares character matrix content for parsing.
}
