% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{text2sentences}
\alias{strsplit2}
\alias{grep2}
\alias{text2num}
\alias{letter.convert}
\alias{pCheck}
\alias{standardStats}
\alias{ngram}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{JATSdecoder}{\code{\link[JATSdecoder]{grep2}}, \code{\link[JATSdecoder]{letter.convert}}, \code{\link[JATSdecoder]{ngram}}, \code{\link[JATSdecoder]{pCheck}}, \code{\link[JATSdecoder]{standardStats}}, \code{\link[JATSdecoder]{strsplit2}}, \code{\link[JATSdecoder]{text2num}}, \code{\link[JATSdecoder]{text2sentences}}}
}}

