% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2stats.R
\name{table2stats}
\alias{table2stats}
\title{table2stats}
\usage{
table2stats(
  x,
  standardPcoding = FALSE,
  noSign2p = TRUE,
  correctComma = FALSE,
  rotate = FALSE,
  expandAbbreviations = TRUE,
  superscript2bracket = TRUE,
  stats.mode = "all",
  checkP = FALSE,
  alpha = 0.05,
  criticalDif = 0.02,
  alternative = "undirected",
  estimateZ = FALSE,
  T2t = FALSE,
  addDF = TRUE,
  collapse = TRUE,
  addTableName = FALSE,
  rm.na.col = TRUE
)
}
\arguments{
\item{x}{Input. Either a file path to an XML, HTML, HML, DOCX, or PDF file; or a matrix object; or a vector of plain HTML-coded tables.}

\item{standardPcoding}{Logical. If TRUE, and no other detection of coding is detected, then standard coding of p-values is assumed to be * p<.05, ** p<.01, and ***p<.001.}

\item{noSign2p}{Logical. If TRUE, imputes 'p>maximum of coded p-values' to cells that are not coded to be significant.}

\item{correctComma}{Logical. If TRUE, decimal sign commas are converted to dots.}

\item{rotate}{Logical. If TRUE, matrix content is parsed by column.}

\item{expandAbbreviations}{Logical. If TRUE, detected abbreviations are expanded to label from table caption/footer.}

\item{superscript2bracket}{Logical. If TRUE, detected superscript codings are inserted inside parentheses.}

\item{stats.mode}{Select a subset of test results by p-value checkability for output. One of: c("all", "checkable", "computable", "uncomputable").}

\item{checkP}{Logical. If TRUE, detected p-values and recalculated p-values will be checked for consistency.}

\item{alpha}{Numeric. Defines the alpha level to be used for error assignment.}

\item{criticalDif}{Numeric. Sets the absolute maximum difference in reported and recalculated p-values for error detection.}

\item{alternative}{Character. Select test sidedness for recomputation of p-values from t-, r-, and beta-values. One of c("undirected", "directed"). If "directed" is specified, p-values for directed null hypotheses are added to the table but still require a manual inspection of the consistency of the direction.}

\item{estimateZ}{Logical. If TRUE, detected beta-/d-values are divided by the reported standard error "SE" to estimate Z-values ("Zest") for observed beta/d and computation of p-values. Note: This is only valid if Gauss-Markov assumptions are met and a sufficiently large sample size is used. If a Z- or t-value is detected in a report of a beta-/d-coefficient with SE, no estimation will be performed, although set to TRUE.}

\item{T2t}{Logical. If TRUE, capital letter T is treated as a t-statistic.}

\item{addDF}{Logical. If TRUE, detected sample size N in the caption/footer is inserted as degrees of freedom (N-2) to r- and t-values that are reported without degrees of freedom.}

\item{collapse}{Logical. If TRUE, the result is collapsed to a single data frame object. Else, a list of data frames with length = n matrices is returned.}

\item{addTableName}{Logical. If TRUE, the table number is added in front of the extracted results.}

\item{rm.na.col}{Logical. If TRUE, removes all columns with only NA.}
}
\value{
A data.frame object with the extracted statistical standard results, recalculated p-values and a rudimentary, optional consistency check for reported p-values (if 'checkP=TRUE').
}
\description{
Extracts tabulated statistical results from documents in XML, HTML, HML, DOCX, or PDF format.
}
