% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2text.R
\name{table2text}
\alias{table2text}
\title{table2text}
\usage{
table2text(
  x,
  unifyMatrix = TRUE,
  unifyStats = FALSE,
  expandAbbreviations = TRUE,
  standardPcoding = FALSE,
  addTableName = TRUE
)
}
\arguments{
\item{x}{A vector with HTML tables or a single file path to an HTML, XML, CERMXML, HML, PDF or DOCX file..}

\item{unifyMatrix}{Logical. If TRUE, matrix cells are unifiedfor better post processing.}

\item{unifyStats}{Logical. If TRUE, output is unified for better post processing (e.g.: "p-value"->"p").}

\item{expandAbbreviations}{Logical. If TRUE, detected abbreviations are expanded to label from table caption/footer.}

\item{standardPcoding}{Logical. If TRUE, and no other detection of coding is detected, then standard coding of p-values is assumed to be * p<.05, ** p<.01 and ***p<.001.}

\item{addTableName}{Logical. If TRUE, table number is added before the parsed text lines.}
}
\value{
A List with parsed table content per HTML table. The result vector in each list element can be further processed with standardStats() to extract and structure the statistical standard test results only.
}
\description{
Parses tabled content from HTML coded content or HTML, DOCX or PDF file to text.
}
