% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx2matrix.R
\name{docx2matrix}
\alias{docx2matrix}
\title{docx2matrix}
\usage{
docx2matrix(x, replicate = TRUE)
}
\arguments{
\item{x}{File path of a DOCX input file.}

\item{replicate}{Logical. If TRUE, replicates content when splitting connected cells.}
}
\value{
List with extracted matrices.
}
\description{
Extracts tables from DOCX documents and return list of character matrices.
}
