% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx2matrix.R
\name{docx2matrix}
\alias{docx2matrix}
\title{docx2matrix}
\usage{
docx2matrix(x, replicate = TRUE)
}
\arguments{
\item{x}{File path to a DOCX input file with tables.}

\item{replicate}{Logical. If TRUE, replicates content when splitting connected cells.}
}
\value{
List with extracted tables as character matrices.
}
\description{
Extracts tables from DOCX documents and returns a list of character matrices.
}
