% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.HTML.tables.R
\name{get.HTML.tables}
\alias{get.HTML.tables}
\title{get.HTML.tables}
\usage{
get.HTML.tables(x)
}
\arguments{
\item{x}{HTML, HML, XML, or CERMXML file or character object with HTML-encoded content.}
}
\value{
Character vector with one HTML-encoded table per cell.
}
\description{
Extracts HTML tables as a vector of HTML-coded tables from plain HTML code, HTML, HML, XML, or CERMXML files.
}
