% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legendCodings.R
\name{legendCodings}
\alias{legendCodings}
\title{legendCodings}
\usage{
legendCodings(x)
}
\arguments{
\item{x}{An HTML-coded table or plain textual input of table caption and/or footnote text.}
}
\value{
A list with detected p-value and superscript codings, abbreviations, and reported sample size/s.
}
\description{
Extracts the coding of p-values, brackets, abbreviations, superscripts, diagonal content, and the reported sample size/s with 'N=number' from table captions and footnote text.
}
\examples{
x<-"+ p>.05, ^**p<.01, SSq, Sum of Squares, ^a t-test, n=120. 
POS: perceived organizational support, JP; job performance.
Numbers in parenthesis are standard errors. 
Bold values indicate significance at p<.05."
legendCodings(x)
}
