% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMatrixContent.R
\name{parseMatrixContent}
\alias{parseMatrixContent}
\title{parseMatrixContent}
\usage{
parseMatrixContent(
  x,
  legend = NULL,
  standardPcoding = TRUE,
  noSign2p = TRUE,
  forceClass = NULL,
  expandAbbreviations = TRUE,
  superscript2bracket = FALSE,
  addDF = TRUE
)
}
\arguments{
\item{x}{A character matrix or list with a character matrix as first and only element.}

\item{legend}{The table's caption/footer notes as a character vector.}

\item{standardPcoding}{Logical. If TRUE, and no other detection of p-value coding is detected, standard coding of p-values is assumed to be: * p<.05, ** p<.01, and *** p<.001.}

\item{noSign2p}{Logical. If TRUE, imputes 'p>maximum of coded p-values' to cells that are not coded to be significant.}

\item{forceClass}{Character. Set a fixed table class for extraction heuristic. One of c("tabled result", "correlation", "matrix", "text").}

\item{expandAbbreviations}{Logical. If TRUE, detected abbreviations are expanded to label detected in table caption/footer with tableParser::legendCodings().}

\item{superscript2bracket}{Logical. If TRUE, detected superscript codings are inserted inside parentheses.}

\item{addDF}{Logical. If TRUE, detected sample size N in the caption/footer is inserted as degrees of freedom (N-2) to r- and t-values that are reported without degrees of freedom.}
}
\value{
A text vector with the parsed matrix content.
}
\description{
Parses character matrix content into a text vector. This is the basic function of tableParser, which is implemented in matrix2text(), table2text(), and table2stats(). Row and column names are parsed to cell content with operators that depend on the cell content. Numeric cells are parsed with "=", and textual cell content with ":". Cells that start with an operator ('<', '=' or '>') are parsed without a separator. Detected codings for (e.g., p-values, abbreviations) from table legend text can be used to extend the tabled content to a fully written-out form.
}
