% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{prepareMatrix}
\usage{
prepareMatrix(x, split = FALSE, forceClass = NULL, na.rm = TRUE, legend = NULL)
}
\arguments{
\item{x}{character matrix}

\item{split}{Logical. If TRUE, multi-model matrices are split into a list of single-model matrices.}

\item{forceClass}{Character. Set matrix-specific handling to one of c("tabled result", "correlation", "matrix, "text").}

\item{na.rm}{Logical. If TRUE, NA cells are set to empty cells.}

\item{legend}{Character. Optional, text from table caption and/or footnote for table class specific processing.}
}
\value{
A character matrix
}
\description{
Prepares character matrix content for parsing. Removes empty rows and columns, extends content from plausible grouping cells to sparse cells, collapses multiple header rows, and splits multiple model tables to a list of single model tables.
}
\examples{
# example matrix
x<-cbind(c("","","name","","",""), 
         c("group","name","A","B","","C"), 
         c("value","","1","2","","3"))
x

# apply function
prepareMatrix(x)
}
