% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMatrix.R
\name{prepareMatrix}
\alias{prepareMatrix}
\title{prepareMatrix}
\usage{
prepareMatrix(x, split = FALSE, forceClass = NULL, na.rm = TRUE)
}
\arguments{
\item{x}{character matrix}

\item{split}{logical. If TRUE, multi-model matrices are split into a list of single-model matrices.}

\item{forceClass}{character. Set matrix-specific handling to one of c("tabled result", "correlation", "matrix, "text").}

\item{na.rm}{Logical. If TRUE, NA cells are set to empty cells.}
}
\value{
A character matrix
}
\description{
Prepares character matrix content for parsing.
}
