% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2text.R
\name{table2text}
\alias{table2text}
\title{table2text}
\usage{
table2text(
  x,
  unifyMatrix = TRUE,
  unifyStats = FALSE,
  expandAbbreviations = TRUE,
  superscript2bracket = TRUE,
  standardPcoding = FALSE,
  addDF = TRUE,
  rotate = FALSE,
  correctComma = FALSE,
  addDescription = TRUE,
  unlist = FALSE,
  addTableName = TRUE
)
}
\arguments{
\item{x}{A vector with HTML tables or a single file path to an HTML, XML, CERMXML, HML, PDF or DOCX file.}

\item{unifyMatrix}{Logical. If TRUE, matrix cells are unified for better post processing.}

\item{unifyStats}{Logical. If TRUE, output is unified for better post processing (e.g.: "p-value"->"p").}

\item{expandAbbreviations}{Logical. If TRUE, detected abbreviations are expanded to label from table caption/footer.}

\item{superscript2bracket}{Logical. If TRUE, detected superscript codings are inserted inside parentheses.}

\item{standardPcoding}{Logical. If TRUE, and no other detection of coding is detected, standard coding of p-values is assumed to be * p<.05, ** p<.01 and ***p<.001.}

\item{addDF}{Logical. If TRUE, detected sample size N in caption/footer is inserted as degrees of freedom (N-2) to r- and t-values that are reported without degrees of freedom.}

\item{rotate}{Logical. If TRUE, matrix content is parsed by column.}

\item{correctComma}{Logical. If TRUE and unifyMatrix=TRUE, decimal sign commas are converted to dots.}

\item{addDescription}{Logical. If TRUE, table caption and footer are added before the extracted table content for better readability.}

\item{unlist}{Logical. If TRUE, output is returned as vector.}

\item{addTableName}{Logical. If TRUE and unlist=TRUE, table number is added in front of unlisted text lines.}
}
\value{
List with parsed tabled content as elements. The text vector in each list element can be further processed with JATSdecoder::standardStats() to extract and structure the statistical standard test results.
}
\description{
Parses tabled content from HTML coded content or HTML, DOCX or PDF file to human readable text vector. Before parsing, header lines are collapsed and connected cells are broken up.
}
