% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClass.R
\name{tableClass}
\alias{tableClass}
\title{tableClass
Classifies matrix content to either 'tabled content', 'correlation', or 'text'}
\usage{
tableClass(x, legend = NULL)
}
\arguments{
\item{x}{A character matrix}

\item{legend}{A text string from tables caption and/or footer}
}
\value{
A character object with the tables class.
}
\description{
tableClass
Classifies matrix content to either 'tabled content', 'correlation', or 'text'
}
