% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableClass.R
\name{tableClass}
\alias{tableClass}
\title{tableClass}
\usage{
tableClass(x, legend = NULL)
}
\arguments{
\item{x}{A character matrix}

\item{legend}{A text vector with the tables caption and/or footnote.}
}
\value{
A character object of length=1 with the table's class.
}
\description{
Classifies matrix content to either 'tabled results', 'correlation', 'matrix', 'text', 'vector', 'model with model statistics', or 'multi model with model statistics'.
}
