% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyMatrixContent.R
\name{unifyMatrixContent}
\alias{unifyMatrixContent}
\title{unifyMatrixContent
Unifies content of character matrices. E.g.: comas in big numbers and HTML tags are removed. Performs space corrections and unifies hyphens and spaces.}
\usage{
unifyMatrixContent(
  x,
  letter.convert = TRUE,
  greek2text = TRUE,
  text2num = TRUE
)
}
\arguments{
\item{x}{a character matrix.}

\item{letter.convert}{Logical. If TRUE hex codes will be unified and converted to unicode with JATSdecoder::letter.convert().}

\item{greek2text}{Logical. If TRUE and 'letter.convert=TRUE', converts and unifies various Greek letters to a text based form (e.g. 'alpha', 'beta').}

\item{text2num}{Logical. If TRUE, textual representations of numbers (words, exponents, fractions) are converted to digit numbers.}
}
\description{
unifyMatrixContent
Unifies content of character matrices. E.g.: comas in big numbers and HTML tags are removed. Performs space corrections and unifies hyphens and spaces.
}
