% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unifyStats.R
\name{unifyStats}
\alias{unifyStats}
\title{unifyStats
Unifies textual representations of statististical results.}
\usage{
unifyStats(x)
}
\arguments{
\item{x}{A text string as vector.}
}
\value{
A unified text string.
}
\description{
unifyStats
Unifies textual representations of statististical results.
}
